Move initialization functions to clkInit.cpp
