+++ CMakeLists.txt
@@ -38,6 +38,7 @@ add_library(recombo_core STATIC
         src/clkTopology.cpp
         src/clkGeometry.cpp
         src/clkConfig.cpp
+        src/clkInit.cpp
         src/mmc_config.cpp
         src/mmchain.cpp
         src/newsud.cpp
